Refatora script em módulos
