+++ index.html
@@ -38,6 +38,6 @@
             <button type="submit">Adicionar</button>
         </form>
     </div>
-    <script src="script.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 </html>
